+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.7.1
+
+### Added
+
+- Adds support for additional source vocabularies to the `rxnorm` sub-command using the `-v` option
+- Adds support for additional suppress flag filtering to the `rxnorm` sub-command using the `-s` option
+
 ## 0.7.0
 
-## Added
+### Added
 
 - Adds a new sub-command: `gsea`